[Auditbeat] Read formula path from INSTALL_RECEIPT.json for Homebrew packages (#10507)
When reading Homebrew packages, changes to reading the path to the Ruby formula from `/usr/local/Cellar/{pkg.Name}/{pkg.Version}/INSTALL_RECEIPT.json` and fall back to the previous `/usr/local/Cellar/{pkg.Name}/{pkg.Version}/.brew/{pkg.Name}.rb` only if that does not work.

(cherry picked from commit 912ef57d864cc6024e0d46151e52f478ec26f6bd)
